Added widgets and icons to the mobile app menu
