+++ homesense/pubspec.yaml
@@ -30,10 +30,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
 
 dev_dependencies:
@@ -60,10 +56,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-  -lib/icons/
-
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+   - lib/icons/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
